refactor: finish Phase 4 module hygiene (#1376)
Complete remaining structural rewrite acceptance criteria:

- Rename ast/extract → extract_to_file; deprecated extract shim (#1376)
- Split symbols into mod + tests; deprecate rewrite re-exports with issue link
- Split plan::Operation into plan/operation.rs (modules under 800-line budget)
- Shared parse_group_position tests; tx uses extract_to_file path
- size-waiver comments on remaining >1k production files
- AGENTS.md AST naming + waiver policy
- Integration tests: module_hygiene_tests (waivers, naming, parsers, deprecations)

Closes #1376
Closes #1372
